Minor nits
